refactor: CafegoryAccessToken 클래스 이름 JwtAccessToken 으로 변경 및 패키지 구조 변경
