+++ src/main/java/com/example/demo/service/auth/JwtService.java
@@ -7,7 +7,7 @@ import java.util.Map;
 
 import com.example.demo.implement.tokenmanagerment.JwtCafegoryTokenManager;
 import com.example.demo.implement.tokenmanagerment.JwtTokenManager;
-import com.example.demo.dto.auth.CafegoryAccessToken;
+import com.example.demo.implement.token.JwtAccessToken;
 import com.example.demo.implement.token.JwtClaims;
 import com.example.demo.exception.ExceptionType;
 import com.example.demo.exception.JwtCustomException;
@@ -37,7 +37,7 @@ public class JwtService {
         );
     }
 
-    public CafegoryAccessToken verifyAndRefreshAccessToken(final String accessToken, final String refreshToken) {
+    public JwtAccessToken verifyAndRefreshAccessToken(final String accessToken, final String refreshToken) {
         //TODO 토큰 재발급 API는 토큰 검증 인터셉터를 거치면 안된다. 토큰 검증 인터셉터는 액세스 토큰의 만료를 검증한다.
         validateNullToken(accessToken, ExceptionType.JWT_ACCESS_TOKEN_MISSING);
         validateNullToken(refreshToken, ExceptionType.JWT_REFRESH_TOKEN_MISSING);
@@ -56,7 +56,7 @@ public class JwtService {
         String newAccessToken = jwtCafegoryTokenManager.createAccessToken(
                 convertMemberToMap(memberInDb)
         );
-        return new CafegoryAccessToken(newAccessToken);
+        return new JwtAccessToken(newAccessToken);
     }
 
     private void validateNullToken(final String token, ExceptionType exceptionType) {
